+++ packages/ui/src/storybook/index.ts
@@ -1,2 +1,2 @@
 // Design System Exports
-export * from "./design-system/DesignTokens";
+export * from "./design-system/02_DesignTokens";