+++ platforms/web/apps/storybook/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @astudio/storybook
+# @design-studio/storybook
 
 Last updated: 2026-01-09
 
@@ -20,5 +20,5 @@ Last updated: 2026-01-09
 - Updated dependencies [b716ca9]
 - Updated dependencies
   - @design-studio/ui@0.1.0
-  - @astudio/tokens@0.1.0
-  - @astudio/runtime@0.0.1
+  - @design-studio/tokens@0.1.0
+  - @design-studio/runtime@0.0.1