+++ platforms/web/apps/web/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @astudio/web
+# @design-studio/web
 
 Last updated: 2026-01-09
 
@@ -20,4 +20,4 @@ Last updated: 2026-01-09
 - Updated dependencies [b716ca9]
 - Updated dependencies
   - @design-studio/ui@0.1.0
-  - @astudio/runtime@0.0.1
+  - @design-studio/runtime@0.0.1